W-15854291: Refactored code to collapse following hierarchy: 1. AbstractFileCommand, ExternalFileCommand, FileCommand - FtpCommand 2. FileSystemProvider and FtpAbstractConnectionProvider -  FtpConnectionProvider
